Beam tfrecords writer. Issue #677. Issue #737.
Create a new version of all Beam datasets, using S3 compatible tfrecords writer.
Default versions of beam datasets are not changed, but cannot generate the datasets at head anymore.

Builders are made pickable, which makes it easier to write Beam operations.

Mechanism to order examples is the same as in non-Beam pipeline:
 1- Hash the keys,
 2- Distributes examples in buckets based on hash(key),
 3- Sort each bucket,
 4- Write final tfrecord shards.

The main difference is we don't handle the files to store the buckets ourselves, Beam does.

As in non-Beam pipeline, not knowing in advance the total size of a split doesn't make it easy to pick the right number of buckets. Went with 100K buckets, which should be fine for splits up to 1PB.

PiperOrigin-RevId: 274156411